requirements.txt
certifi==2023.7.22
charset-normalizer==2.1.0
Django==2.2.28
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-storages==1.8
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==3.3
numpy==1.23.1
psycopg2-binary==2.8.6
PyJWT==2.4.0
python-decouple==3.1
pytz==2018.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==2.63.0
urllib3==1.26.11